feat: replace native time picker with custom dropdown for strict 15-min enforcement

+++ backend/src/main/java/com/medical/wiki/dto/PersonalDashboardDto.java
@@ -13,7 +13,7 @@ public class PersonalDashboardDto {
 
     private int pendingLeaveRequestsCount;
     private int approvedLeaveRequestsCount;
-    private int paidLeaveDays;
+    private double paidLeaveDays;
 
     private int unreadNotificationsCount;
 }
